Greatly simplify test suite
Mainly the change is that it won't build the latest release to compare against the current master branch. It just grabs the artifact path and simply tests that. This makes it a lot faster and a lot simpler, the comparison was mostly unnecessary anyway.

+++ test/pixi_test.sh
@@ -26,52 +26,13 @@ else
     exit 1
 fi
 
-if [[ "$OSTYPE" == "darwin"* ]]; then
-    asar_shared_name="libasar.dylib"
-else
-    asar_shared_name="libasar.so"
-fi
-
 wget www.atarismwc.com/base.smc
-
-# build latest pixi release along with zip and move in main directory
-curl -s https://api.github.com/repos/JackTheSpades/SpriteToolSuperDelux/releases/latest | grep "tarball_url" | cut -d '"' -f 4 | wget -O latest.tar.gz -qi -
-mkdir latest && tar xf latest.tar.gz -C latest --strip-components 1
-rm latest.tar.gz
-cp zip_pixi_rasp.py latest/zip_pixi_rasp.py
-cd latest
-spritecontents=$(cat src/sprite.cpp)
-line=$'#include <vector>\n'
-echo "$line$spritecontents" > src/sprite.cpp
-if [[ "$OSTYPE" == "darwin"* ]]; then
-    sed -i '' s/"-Wl,--gc-sections"//g Makefile
-fi
-make
-mkdir build
-mv pixi build/pixi
-cd ..
-
-# build asar 1.81
-curl -s https://api.github.com/repos/RPGHacker/asar/releases/latest | grep "tarball_url" | cut -d '"' -f 4 | wget -O asar_latest.tar.gz -qi -
-mkdir asar_latest && tar xf asar_latest.tar.gz -C asar_latest --strip-components 1
-rm asar_latest.tar.gz
-cd asar_latest
-cmake src && make
-cd ..
-
-# move libasar in latest and build zip
-cp asar_latest/asar/$asar_shared_name latest/$asar_shared_name
-cd latest
-python3 zip_pixi_rasp.py
-mv pixi.zip ../pixi_latest.zip
-cd ..
 
 if [[ -z "${ARTIFACT_PATH}" ]]; then
     # clone current pixi repo and build zip
     CC=gcc-11
     CXX=g++-11
     git clone $repourl
-    cp asar_latest/asar/$asar_shared_name SpriteToolSuperDelux/$asar_shared_name
     cd SpriteToolSuperDelux
     git checkout $branch
     mkdir build
@@ -85,38 +46,32 @@ else
     cp "${ARTIFACT_PATH}" pixi.zip
 fi
 
-# unzip both pixi zips (latest stable and current)
 unzip pixi.zip -d pixi
-unzip pixi_latest.zip -d pixi_latest
 
 # delete temp files
-rm -rf latest
-rm -rf asar_latest
-rm -rf SpriteToolSuperDelux
+if [[ -z "${ARTIFACT_PATH}" ]]; then
+    rm -rf SpriteToolSuperDelux
+fi
 rm pixi.zip
-rm pixi_latest.zip
 
 mkdir downloader_test
-cp downloader.py downloader_test/downloader.py
+cp runner.py downloader_test/runner.py
 
 # move rom and start script
 cp -r pixi downloader_test/pixi
-mv downloader_test/pixi/pixi downloader_test/pixi/pixi_current
-cp pixi_latest/pixi downloader_test/pixi/pixi_latest
-cp base.smc downloader_test/pixi/base_current.smc
-cp base.smc downloader_test/pixi/base_latest.smc
+cp base.smc downloader_test/pixi/base.smc
 if [ -d ".sprites_dl_cache" ]; then
     cp -r .sprites_dl_cache/* downloader_test
     cd downloader_test 
-    yes | python3 downloader.py "true"
+    yes | python3 runner.py "true"
     cd ..
 else
     mkdir .sprites_dl_cache
     cd downloader_test
     if [[ "$OSTYPE" != "darwin"* ]]; then
-        yes | python3 downloader.py "false"
+        yes | python3 runner.py "false"
     else
-        python3 downloader.py "false"
+        python3 runner.py "false"
     fi
     echo "Finished testing all sprites"
     cd ..
@@ -130,22 +85,9 @@ else
 fi
 
 echo "Moving results"
-mv downloader_test/result_current.json result_current.json
-mv downloader_test/result_latest.json result_latest.json
-mv downloader_test/differences.json differences.json
+mv downloader_test/result.json result.json
 
 echo "Deleting temp files"
 rm -rf downloader_test
 rm -rf pixi
-rm -rf pixi_latest
-
-# we have the json with the results
-echo "Producing the result json"
-res=$(python3 check_diff.py)
-
-# final clean up
-echo "Final cleanup"
-rm result_current.json
-rm result_latest.json
-rm differences.json
 rm base.smc
